app.js moved to components folder
