+++ src/components/App.js
@@ -1,11 +1,9 @@
 import React from 'react';
-import logo from './logo.svg';
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
         <h1>
           BagaJob
         </h1>
